+++ src/services/answer-refiner/answer-refiner.ts
@@ -1,13 +1,12 @@
-import type {
+import {
   AnswerData,
   BatchRefinementOptions,
   BatchRefinementResult,
-  GeminiConfig,
   RefinementResult,
   WriteResult,
-} from "../../types";
+} from "./types";
+import { GeminiClient, GeminiConfig } from "@/services/gemini-client";
 
-import { GeminiClient } from "@/services/gemini-client";
 import { GoogleServiceFactory } from "@/services/google-service-factory";
 import { GoogleSheetsService } from "@/services/google-sheets";
