+++ src/services/feedback-generator/index.ts
@@ -1,1 +1,2 @@
 export * from "./feedback-generator";
+export * from "./types";
