+++ src/services/gemini-client/gemini-client.ts
@@ -1,8 +1,8 @@
-import type {
+import {
   GeminiConfig,
   TextProcessingRequest,
   TextProcessingResponse,
-} from "@/types";
+} from "./types";
 
 import { GoogleGenAI } from "@google/genai";
