+++ src/services/gemini-client/index.ts
@@ -1,1 +1,2 @@
 export * from "./gemini-client";
+export * from "./types";
