+++ src/services/google-auth/index.ts
@@ -1,1 +1,2 @@
 export * from "./google-auth";
+export * from "./types";
