+++ src/services/google-auth/types/index.ts
@@ -1,0 +1,5 @@
+export interface AuthResult {
+  success: boolean;
+  token?: string;
+  error?: string;
+}
