+++ src/services/google-sheets/google-sheets.ts
@@ -1,34 +1,11 @@
+import type {
+  ApiResult,
+  AppendResult,
+  RangeData,
+  SpreadsheetMetadata,
+} from "./types";
+
 import type { GoogleAuthService } from "@/services/google-auth";
-
-export interface ApiResult<T = unknown> {
-  success: boolean;
-  data?: T;
-  error?: string;
-}
-
-export interface SpreadsheetMetadata {
-  spreadsheetId: string;
-  properties: {
-    title: string;
-  };
-  sheets: Array<{
-    properties: {
-      title: string;
-    };
-  }>;
-}
-
-export interface RangeData {
-  range: string;
-  values?: string[][];
-}
-
-export interface AppendResult {
-  spreadsheetId: string;
-  updates: {
-    updatedRows: number;
-  };
-}
 
 export class GoogleSheetsService {
   private readonly baseUrl = "https://sheets.googleapis.com/v4/spreadsheets";
