+++ src/services/google-sheets/index.ts
@@ -1,1 +1,2 @@
 export * from "./google-sheets";
+export * from "./types";
