+++ src/services/index.ts
@@ -3,4 +3,5 @@ export { FeedbackGenerator } from "./feedback-generator";
 export { GeminiClient } from "./gemini-client";
 export { GoogleAuthService } from "./google-auth";
 export { GoogleSheetsService } from "./google-sheets";
-export { GoogleServiceFactory as GoogleSheetsServiceFactory } from "./google-service-factory";
+export { GoogleServiceFactory } from "./google-service-factory";
+export { SheetDetector } from "./sheet-detector";
