+++ src/services/sheet-detector/index.ts
@@ -1,1 +1,2 @@
 export * from "./sheet-detector";
+export * from "./types";
