+++ src/services/sheet-detector/types/index.ts
@@ -1,0 +1,5 @@
+export interface SheetInfo {
+  spreadsheetId: string | null;
+  sheetName: string;
+  url: string;
+}
